Added sleep for service readiness; added case where container is missing handling

+++ src/_podman.sh
@@ -355,11 +355,19 @@ function podman_start_cke() {
   local env_file="--env-file ${PLEXTRAC_HOME:-}/.env"
   local env_vars="-e DATABASE_DATABASE=${PG_CKEDITOR_DB:?err} -e DATABASE_DRIVER=postgres -e DATABASE_HOST=postgres -e DATABASE_PASSWORD=${PG_CKEDITOR_ADMIN_PASSWORD:?err} -e DATABASE_POOL_CONNECTION_LIMIT=10 -e DATABASE_PORT=5432 -e DATABASE_SCHEMA=public -e DATABASE_USER=${PG_CKEDITOR_ADMIN_USER:?err} -e ENABLE_METRIC_LOGS=${CKEDITOR_ENABLE_METRIC_LOGS:-false} -e ENVIRONMENTS_MANAGEMENT_SECRET_KEY=${CKEDITOR_ENVIRONMENT_SECRET_KEY:-} -e LICENSE_KEY=${CKEDITOR_SERVER_LICENSE_KEY:-} -e LOG_LEVEL=${CKEDITOR_LOG_LEVEL:-} -e REDIS_CONNECTION_STRING=redis://redis:6379 -e REDIS_HOST=redis -e REDIS_PASSWORD=${REDIS_PASSWORD:?err}"
 
-  podman stop ckeditor-backend 1>/dev/null
-  podman rm -f ckeditor-backend 1>/dev/null
-  podman image prune -f 1>/dev/null
+  debug "podman_start_cke: Checking if ckeditor-backend exists then stopping and removing"
+  if ! podman container exists ckeditor-backend; then
+    debug "podman_start_cke: ckeditor-backend does not exist, skipping stop and remove"
+  else
+    debug "podman_start_cke: ckeditor-backend exists, stopping and removing"
+    podman stop ckeditor-backend 1>/dev/null
+    podman rm -f ckeditor-backend 1>/dev/null
+    podman image prune -f 1>/dev/null
+  fi
   debug "podman_start_cke: Starting CKEditor Backend"
   container_client run $env_file $env_vars $init $alias $entrypoint $restart_policy $healthcheck \
         $volumes --name=ckeditor-backend $deploy --network=plextrac $ports -d $image 1>/dev/null
-  debug "podman_start_cke: CKEditor Backend started"
+  debug "podman_start_cke: CKEditor Backend started, waiting 10 seconds for it to initialize"
+  info "Waiting 10 seconds for CKEditor Backend to initialize"
+  sleep 10
 }
